pyproject.toml
[tool.poetry]
name = "rubix-mapping"
version = "1.0.0"
description = "Rubix Mapping Service"
authors = ["NubeIO <tech-support@nube-io.com>"]

[tool.poetry.dependencies]
python = "^3.6"
gunicorn = "^20.0.4"
click = "^7.1.2"
gevent = [
    { version = "20.12.1", markers = "platform_machine == 'x86_64'" },
    { git = "https://github.com/gevent/gevent.git", tag = "20.12.1", markers = "platform_machine != 'x86_64'" }
]
rubix-http = {git = "https://github.com/NubeIO/rubix-http", rev = "v1.1.1"}
Flask-Migrate = "2.6.0"
Flask-Script = "2.0.6"
shortuuid = "1.0.1"
rubix-mqtt = {git = "https://github.com/NubeIO/rubix-mqtt", rev = "v1.2.0"}
rubix-registry = {git = "https://github.com/NubeIO/rubix-registry", rev = "v1.1.1"}

[tool.poetry.dev-dependencies]
pyinstaller = "^4.1"
pytest = "^6.2.2"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
